+++ activerecord/lib/active_record/connection_adapters/abstract_adapter.rb
@@ -238,6 +238,11 @@ module ActiveRecord
         false
       end
 
+      # Does this adapter support expression indices?
+      def supports_expression_index?
+        false
+      end
+
       # Does this adapter support explain?
       def supports_explain?
         false