+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -139,6 +139,10 @@ module ActiveRecord
         true
       end
 
+      def supports_expression_index?
+        true
+      end
+
       def supports_transaction_isolation?
         true
       end